refactor: replace float with BigDecimal for monetary fields
Entities (Product, QuoteItem, Quote), DTOs (ProductRequest/Response,
QuoteItemResponse, QuoteResponse), and services updated. float[] column
widths in PdfServiceImpl unchanged (iText layout, not monetary).

+++ src/main/java/com/slim/kreadevis_backend/dto/product/ProductResponse.java
@@ -1,12 +1,14 @@
 package com.slim.kreadevis_backend.dto.product;
 
+import java.math.BigDecimal;
+
 public record ProductResponse(
         Long id,
         String label,
         String description,
         Long stockQuantity,
-        float unitPrice,
-        float vatRate,
+        BigDecimal unitPrice,
+        BigDecimal vatRate,
         String referenceCode,
         Long supplierId
 ) {}
