+++ src/main/java/com/slim/kreadevis_backend/dto/quote/QuoteResponse.java
@@ -3,6 +3,7 @@ package com.slim.kreadevis_backend.dto.quote;
 import com.slim.kreadevis_backend.dto.client.ClientResponse;
 import com.slim.kreadevis_backend.entity.QuoteStatus;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -10,7 +11,7 @@ public record QuoteResponse(
         Long id,
         String referenceCode,
         LocalDate date,
-        float totalPrice,
+        BigDecimal totalPrice,
         QuoteStatus status,
         ClientResponse client,
         List<QuoteItemResponse> items
